requirements.txt
opencv_python==4.0.0.21
Keras==2.2.4
numpy==1.16.0
pandas==0.23.4
matplotlib==3.0.2
scikit_image==0.14.2
Pillow==5.4.1
skimage==0.0
scikit_learn==0.20.2
